# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CtxOS API
  description: Context-driven OS Intelligence and Analysis API
  version: 1.0.0
  contact:
    name: CtxOS Team
    email: api@ctxos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ctxos.com
    description: Production server
...
paths:
...
  /api/v1/score:
    post:
      tags: [scoring]
      summary: Score entity
      description: Score an entity using specified scoring engines
      operationId: scoreEntity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRequest'
      responses:
        '200':
          description: Entity scored successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoringResult'
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions
        '422':
          description: Validation error
...
  /api/v1/config:
    get:
      tags: [config]
      summary: Get configuration
      description: Get all configuration values
      operationId: getConfig
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuration retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityType:
      type: string
      enum: [host, domain, ip, user, service, application, network, database]
    SignalType:
      type: string
      enum: [vulnerability, port, service, dns, ssl_certificate, configuration, activity, authentication, dependency, subdomain, whois]
    SeverityLevel:
      type: string
      enum: [critical, high, medium, low, info]
    Context:
      type: object
      required: [entity, signals]
      properties:
        entity:
          $ref: '#/components/schemas/Entity'
        signals:
          type: array
          items:
            $ref: '#/components/schemas/Signal'
    Entity:
      type: object
      required: [id, entity_type, name]
      properties:
        id:
          type: string
          description: Entity ID
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Entity type
        name:
          type: string
          description: Entity name
        description:
          type: string
          description: Entity description
        properties:
          type: object
          description: Entity properties
    Signal:
      type: object
      required: [id, source, signal_type, severity, description]
      properties:
        id:
          type: string
          description: Signal ID
        source:
          type: string
          description: Signal source
        signal_type:
          $ref: '#/components/schemas/SignalType'
          description: Signal type
        severity:
          $ref: '#/components/schemas/SeverityLevel'
          description: Signal severity
        description:
          type: string
          description: Signal description
        timestamp:
          type: string
          format: date-time
          description: Signal timestamp
        entity_id:
          type: string
          description: Associated entity ID
        properties:
          type: object
          description: Signal properties

    # Scoring Schemas
    ScoreRequest:
      type: object
      required: [context, engines]
      properties:
        context:
          $ref: '#/components/schemas/Context'
          description: Entity context
        engines:
          type: array
          items:
            type: string
          description: Scoring engines to use
        include_recommendations:
          type: boolean
          default: true
          description: Include recommendations
...
    ScoringResult:
      type: object
      required: [entity_id, entity_type, engine_name, score, severity, timestamp]
      properties:
        entity_id:
          type: string
          description: Entity ID
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Entity type
        engine_name:
          type: string
          description: Scoring engine name
        score:
          type: number
          format: float
          description: Risk score (0-100)
        severity:
          $ref: '#/components/schemas/SeverityLevel'
          description: Risk severity
        factors:
          type: object
          description: Score factors
        recommendations:
          type: array
          items:
            type: string
          description: Recommendations
        timestamp:
          type: string
          format: date-time
          description: Scoring timestamp
...
    ConfigResponse:
      type: object
      required: [config, metadata]
      properties:
        config:
          type: object
          description: Configuration values
        metadata:
          type: object
          description: Configuration metadata
...